Moves knex file back to root
