+++ app/models/playlist.js
@@ -1,6 +1,6 @@
 // define database
 const environment = process.env.NODE_ENV || 'development';
-const configuration = require('../../config/knexfile')[environment];
+const configuration = require('../../knexfile')[environment];
 const database = require('knex')(configuration);
 
 const all = () => database('playlists')
