+++ server.js
@@ -4,7 +4,7 @@ const route = require('./config/routes')
 const bodyParser = require('body-parser');
 
 const environment = process.env.NODE_ENV || 'development';
-const configuration = require('./config/knexfile')[environment];
+const configuration = require('./knexfile')[environment];
 const database = require('knex')(configuration);
 
 app.use(bodyParser.json());
